# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trivia game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create a player by name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a player
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: The user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
...
  /games/{gameId}/questions:
    get:
      operationId: listGameQuestions
      tags: [questions]
      summary: List questions of a game ordered by order_index
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Add a question to a game (admin)
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{questionId}:
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question (admin)
      parameters:
        - name: questionId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Not found
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question (admin)
      parameters:
        - name: questionId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /games/{gameId}/join:
    post:
      operationId: joinGame
      tags: [play]
      summary: Join a game as a participant
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinGameInput"
      responses:
        "201":
          description: Joined (or already joined)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
        "404":
          description: Game not found
  /games/{gameId}/participants:
    get:
      operationId: listGameParticipants
      tags: [play]
      summary: Leaderboard - participants with scores, sorted descending
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Participants with names and scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParticipantWithUser"
  /games/{gameId}/answers:
    post:
      operationId: submitAnswer
      tags: [play]
      summary: Submit an answer to a question
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerInput"
      responses:
        "201":
          description: Answer graded and recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerResult"
        "404":
          description: Game or question not found
        "409":
          description: Already answered this question
  /games/{gameId}/users/{userId}/answers:
    get:
      operationId: listUserAnswers
      tags: [play]
      summary: List a user's answers for a game
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: integer }
        - name: userId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Answers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Answer"
...
components:
  schemas:
...
    UserInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
    User:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        createdAt: { type: string }
...
    QuestionInput:
      type: object
      required: [questionText, questionType, correctAnswer, points, orderIndex]
      properties:
        questionText: { type: string, minLength: 1 }
        questionType:
          type: string
          enum: [multiple_choice, write_in, matching, image_recognition, true_false]
        correctAnswer: { type: string, minLength: 1 }
        options:
          type: ["object", "null"]
          additionalProperties: true
        imageUrl: { type: ["string", "null"] }
        points: { type: integer, minimum: 1 }
        orderIndex: { type: integer, minimum: 0 }
        source: { type: ["string", "null"] }
        factCheckUrl: { type: ["string", "null"] }
        verifiedByAdmin: { type: boolean }
        aiGenerated: { type: boolean }
    QuestionUpdate:
      type: object
      properties:
        questionText: { type: string, minLength: 1 }
        questionType:
          type: string
          enum: [multiple_choice, write_in, matching, image_recognition, true_false]
        correctAnswer: { type: string, minLength: 1 }
        options:
          type: ["object", "null"]
          additionalProperties: true
        imageUrl: { type: ["string", "null"] }
        points: { type: integer, minimum: 1 }
        orderIndex: { type: integer, minimum: 0 }
        source: { type: ["string", "null"] }
        factCheckUrl: { type: ["string", "null"] }
        verifiedByAdmin: { type: boolean }
...
    Question:
      type: object
      required: [id, gameId, questionText, questionType, options, imageUrl, points, orderIndex,verifiedByAdmin, aiGenerated]
      properties:
        id: { type: integer }
        gameId: { type: integer }
        questionText: { type: string }
        questionType:
          type: string
          enum: [multiple_choice, write_in, matching, image_recognition, true_false]
        correctAnswer: { type: string }
        options:
          type: ["object", "null"]
          additionalProperties: true
        imageUrl: { type: ["string", "null"] }
        points: { type: integer }
        orderIndex: { type: integer }
        source: { type: ["string", "null"] }
        factCheckUrl: { type: ["string", "null"] }
        verifiedByAdmin: { type: boolean }
        aiGenerated: { type: boolean }
    JoinGameInput:
      type: object
      additionalProperties: false
    Participant:
      type: object
      required: [id, gameId, userId, totalScore, joinedAt]
      properties:
        id: { type: integer }
        gameId: { type: integer }
        userId: { type: integer }
        totalScore: { type: integer }
        joinedAt: { type: string }
    ParticipantWithUser:
      type: object
      required: [id, gameId, userId, userName, totalScore, joinedAt]
      properties:
        id: { type: integer }
        gameId: { type: integer }
        userId: { type: integer }
        userName: { type: string }
        totalScore: { type: integer }
        joinedAt: { type: string }
    AnswerInput:
      type: object
      required: [questionId, userAnswer]
      properties:
        questionId: { type: integer }
        userAnswer: { type: string }
    Answer:
      type: object
      required: [id, userId, gameId, questionId, userAnswer, isCorrect, pointsEarned,answeredAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        gameId: { type: integer }
        questionId: { type: integer }
        userAnswer: { type: string }
        isCorrect: { type: boolean }
        pointsEarned: { type: integer }
        answeredAt: { type: string }
        correctAnswer: { type: string }
    AnswerResult:
      type: object
      required: [id, userId, gameId, questionId, userAnswer, isCorrect, answeredAt,pointsEarned, totalScore]
      properties:
        id: { type: integer }
        userId: { type: integer }
        gameId: { type: integer }
        questionId: { type: integer }
        userAnswer: { type: string }
        isCorrect: { type: boolean }
        answeredAt: { type: string }
        pointsEarned: { type: integer }
        totalScore: { type: integer }
...